Backend Integration/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
python-magic==0.4.27
aiofiles==23.2.1
tenacity==8.2.3
